Refactored prediction and tree abstractions to be under ml.prediction sub-package
