+++ examples/src/main/scala/org/apache/spark/examples/ml/DecisionTreeExample.scala
@@ -22,7 +22,7 @@ import scala.language.reflectiveCalls
 
 import scopt.OptionParser
 
-import org.apache.spark.ml.tree.DecisionTreeModel
+import org.apache.spark.ml.prediction.tree.DecisionTreeModel
 import org.apache.spark.{SparkConf, SparkContext}
 import org.apache.spark.examples.mllib.AbstractParams
 import org.apache.spark.ml.{Pipeline, PipelineStage}
