+++ mllib/src/main/scala/org/apache/spark/ml/regression/Regressor.scala
@@ -18,7 +18,8 @@
 package org.apache.spark.ml.regression
 
 import org.apache.spark.annotation.{DeveloperApi, AlphaComponent}
-import org.apache.spark.ml.impl.estimator.{PredictionModel, Predictor, PredictorParams}
+import org.apache.spark.ml.impl.estimator.PredictionModel
+import org.apache.spark.ml.prediction.impl.{PredictionModel, PredictorParams, Predictor}
 
 /**
  * :: DeveloperApi ::
