+++ mllib/src/test/scala/org/apache/spark/ml/impl/TreeTests.scala
@@ -24,7 +24,7 @@ import org.scalatest.FunSuite
 import org.apache.spark.api.java.JavaRDD
 import org.apache.spark.ml.attribute.{AttributeGroup, NominalAttribute, NumericAttribute}
 import org.apache.spark.ml.impl.tree._
-import org.apache.spark.ml.tree.{DecisionTreeModel, InternalNode, LeafNode, Node}
+import org.apache.spark.ml.prediction.tree.{DecisionTreeModel, InternalNode, LeafNode, Node}
 import org.apache.spark.mllib.regression.LabeledPoint
 import org.apache.spark.rdd.RDD
 import org.apache.spark.sql.{SQLContext, DataFrame}
